Rewrite People's Choice Vote page logic
